[Feat] #5 - Custom Color Tag View
