+++ Memento-iOS/Memento-iOS/Source/Presentation/Today/View/ColorTagView.swift
@@ -1,0 +1,18 @@
+//
+//  ColorTagView.swift
+//  Memento-iOS
+//
+//  Created by Gahyun Kim on 1/8/25.
+//
+
+import SwiftUI
+
+struct ColorTagView: View {
+    var colorType: String
+    
+    var body: some View {
+        Rectangle()
+            .fill(Color.distinguishColorType(colorType))
+            .frame(width: 3)
+    }
+}
